Pulling changes into fork

+++ settings/js/users/groups.js
@@ -5,17 +5,21 @@
  * See the COPYING-README file.
  */
 
-var $userGroupList;
+var $userGroupList,
+	$sortGroupBy;
 
 var GroupList;
 GroupList = {
 	activeGID: '',
 	everyoneGID: '_everyone',
+	filter: '',
+	filterGroups: false,
 
 	addGroup: function (gid, usercount) {
 		var $li = $userGroupList.find('.isgroup:last-child').clone();
 		$li
 			.data('gid', gid)
+			.attr('data-gid', gid)
 			.find('.groupname').text(gid);
 		GroupList.setUserCount($li, usercount);
 
@@ -27,6 +31,11 @@ GroupList = {
 	},
 
 	setUserCount: function (groupLiElement, usercount) {
+		if ($sortGroupBy !== 1) {
+			// If we don't sort by group count we don't display them either
+			return;
+		}
+
 		var $groupLiElement = $(groupLiElement);
 		if (usercount === undefined || usercount === 0 || usercount < 0) {
 			usercount = '';
@@ -41,18 +50,26 @@ GroupList = {
 		return parseInt($groupLiElement.data('usercount'), 10);
 	},
 
-	modEveryoneCount: function(diff) {
-		var $li = GroupList.getGroupLI(GroupList.everyoneGID);
+	modGroupCount: function(gid, diff) {
+		var $li = GroupList.getGroupLI(gid);
 		var count = GroupList.getUserCount($li) + diff;
 		GroupList.setUserCount($li, count);
 	},
 
 	incEveryoneCount: function() {
-		GroupList.modEveryoneCount(1);
+		GroupList.modGroupCount(GroupList.everyoneGID, 1);
 	},
 
 	decEveryoneCount: function() {
-		GroupList.modEveryoneCount(-1);
+		GroupList.modGroupCount(GroupList.everyoneGID, -1);
+	},
+
+	incGroupCount: function(gid) {
+		GroupList.modGroupCount(gid, 1);
+	},
+
+	decGroupCount: function(gid) {
+		GroupList.modGroupCount(gid, -1);
 	},
 
 	getCurrentGID: function () {
@@ -63,6 +80,33 @@ GroupList = {
 		var lis = $userGroupList.find('.isgroup').get();
 
 		lis.sort(function (a, b) {
+			// "Everyone" always at the top
+			if ($(a).data('gid') === '_everyone') {
+				return -1;
+			} else if ($(b).data('gid') === '_everyone') {
+				return 1;
+			}
+
+			// "admin" always as second
+			if ($(a).data('gid') === 'admin') {
+				return -1;
+			} else if ($(b).data('gid') === 'admin') {
+				return 1;
+			}
+
+			if ($sortGroupBy === 1) {
+				// Sort by user count first
+				var $usersGroupA = $(a).data('usercount'),
+					$usersGroupB = $(b).data('usercount');
+				if ($usersGroupA > 0 && $usersGroupA > $usersGroupB) {
+					return -1;
+				}
+				if ($usersGroupB > 0 && $usersGroupB > $usersGroupA) {
+					return 1;
+				}
+			}
+
+			// Fallback or sort by group name
 			return UserList.alphanum(
 				$(a).find('a span').text(),
 				$(b).find('a span').text()
@@ -93,14 +137,10 @@ GroupList = {
 					var addedGroup = result.groupname;
 					UserList.availableGroups = $.unique($.merge(UserList.availableGroups, [addedGroup]));
 					GroupList.addGroup(result.groupname);
-
-					$('.groupsselect, .subadminsselect')
-						.append($('<option>', { value: result.groupname })
-							.text(result.groupname));
 				}
 				GroupList.toggleAddGroup();
-			}).fail(function(result, textStatus, errorThrown) {
-				OC.dialogs.alert(result.responseJSON.message, t('settings', 'Error creating group'));
+			}).fail(function(result) {
+				OC.Notification.showTemporary(t('settings', 'Error creating group: {message}', {message: result.responseJSON.message}));
 			});
 	},
 
@@ -112,8 +152,9 @@ GroupList = {
 		$.get(
 			OC.generateUrl('/settings/users/groups'),
 			{
-				pattern: filter.getPattern(),
-				filterGroups: filter.filterGroups ? 1 : 0
+				pattern: this.filter,
+				filterGroups: this.filterGroups ? 1 : 0,
+				sortGroups: $sortGroupBy
 			},
 			function (result) {
 
@@ -182,6 +223,7 @@ GroupList = {
 			$('#newgroup-form').show();
 			$('#newgroup-init').hide();
 			$('#newgroupname').focus();
+			GroupList.handleAddGroupInput('');
 		}
 		else {
 			$('#newgroup-form').hide();
@@ -190,11 +232,19 @@ GroupList = {
 		}
 	},
 
+	handleAddGroupInput: function (input) {
+		if(input.length) {
+			$('#newgroup-form input[type="submit"]').attr('disabled', null);
+		} else {
+			$('#newgroup-form input[type="submit"]').attr('disabled', 'disabled');
+		}
+	},
+
 	isGroupNameValid: function (groupname) {
 		if ($.trim(groupname) === '') {
-			OC.dialogs.alert(
-				t('settings', 'A valid group name must be provided'),
-				t('settings', 'Error creating group'));
+			OC.Notification.showTemporary(t('settings', 'Error creating group: {message}', {
+				message: t('settings', 'A valid group name must be provided')
+			}));
 			return false;
 		}
 		return true;
@@ -219,22 +269,25 @@ GroupList = {
 		GroupDeleteHandler = new DeleteHandler('/settings/users/groups', 'groupname',
 			GroupList.hide, GroupList.remove);
 
-		//configure undo
-		OC.Notification.hide();
 		var msg = escapeHTML(t('settings', 'deleted {groupName}', {groupName: '%oid'})) + '<span class="undo">' +
 			escapeHTML(t('settings', 'undo')) + '</span>';
 		GroupDeleteHandler.setNotification(OC.Notification, 'deletegroup', msg,
-			GroupList.show);
+			GroupList.remove);
 
 		//when to mark user for delete
 		$userGroupList.on('click', '.delete', function () {
 			// Call function for handling delete/undo
-			GroupDeleteHandler.mark(GroupList.getElementGID(this));
-		});
-
-		//delete a marked user when leaving the page
-		$(window).on('beforeunload', function () {
-			GroupDeleteHandler.deleteEntry();
+			var groupName = encodeURIComponent(GroupList.getElementGID(this)).toString();
+			OC.dialogs.confirm(
+				t('settings', 'You are about to delete a group. This action can\'t be undone and is permanent. Are you sure that you want to permanently delete {groupName}?', {groupName:decodeURIComponent(groupName)}),
+				t('settings', 'Delete group'),
+				function (confirmation) {
+					if (confirmation) {
+						GroupDeleteHandler.mark(groupName);
+						GroupDeleteHandler.deleteEntry();
+					}
+				}
+			);
 		});
 	},
 
@@ -245,16 +298,16 @@ GroupList = {
 	},
 
 	getElementGID: function (element) {
-		return ($(element).closest('li').data('gid') || '').toString();
+		return ($(element).closest('li').attr('data-gid') || '');
 	},
 	getEveryoneCount: function () {
 		$.ajax({
 			type: "GET",
 			dataType: "json",
-			url: OC.generateUrl('/settings/ajax/geteveryonecount')
+			url: OC.generateUrl('/settings/users/stats')
 		}).success(function (data) {
-			$('#everyonegroup').data('usercount', data.count);
-			$('#everyonecount').text(data.count);
+			$('#everyonegroup').data('usercount', data.totalUsers);
+			$('#everyonecount').text(data.totalUsers);
 		});
 	}
 };
@@ -262,7 +315,11 @@ GroupList = {
 $(document).ready( function () {
 	$userGroupList = $('#usergrouplist');
 	GroupList.initDeleteHandling();
-	GroupList.getEveryoneCount();
+	$sortGroupBy = $userGroupList.data('sort-groups');
+	if ($sortGroupBy === 1) {
+		// Disabled due to performance issues, when we don't need it for sorting
+		GroupList.getEveryoneCount();
+	}
 
 	// Display or hide of Create Group List Element
 	$('#newgroup-form').hide();
@@ -294,5 +351,9 @@ $(document).ready( function () {
 	// click on group name
 	$userGroupList.on('click', '.isgroup', function () {
 		GroupList.showGroup(GroupList.getElementGID(this));
+	});
+
+	$('#newgroupname').on('input', function(){
+		GroupList.handleAddGroupInput(this.value);
 	});
 });
